requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
faiss-cpu
sentence-transformers
openai
pydantic
langchain
langchain_community
langchain_openai
python-dotenv
qdrant_client